Extract identical methods to a base class
Added a base service class and extracted `getLatestBuildId` method
and extended both coverage and tests services from the base class.
Tests should be run for: [AzureDevOpsCoverage] and [AzureDevOpsTests].
